refactor: remove useGame mixin
