+++ package.json
@@ -21,7 +21,6 @@
     "socket.io-client": "^2.3.0",
     "vue": "^2.6.10",
     "vue-router": "^3.1.3",
-    "vue-svg-loader": "^0.16.0",
     "vuetify": "^2.1.0",
     "vuex": "^3.1.2"
   },
@@ -43,6 +42,7 @@
     "sass": "^1.19.0",
     "sass-loader": "^8.0.0",
     "vue-cli-plugin-vuetify": "^2.0.3",
+    "vue-svg-loader": "^0.16.0",
     "vue-template-compiler": "^2.6.10",
     "vuetify-loader": "^1.3.0",
     "webpack": "^4.41.5"
